login

+++ Conectados2/Seguridad/RolPermiso.cs
@@ -6,7 +6,7 @@ namespace Conectados2.Seguridad
     public partial class RolPermiso
     {
         public int IdRolPermiso { get; set; }
-        public long IdRol { get; set; }
+        public int IdRol { get; set; }
         public int IdPermiso { get; set; }
 
         public Permiso IdPermisoNavigation { get; set; }
